added notebook widget to view changes inline
