+++ src/vs/workbench/contrib/notebook/browser/services/notebookEditorService.ts
@@ -5,7 +5,6 @@
 
 import { CodeWindow } from '../../../../../base/browser/window.js';
 import { createDecorator, ServicesAccessor } from '../../../../../platform/instantiation/common/instantiation.js';
-import { NotebookEditorInput } from '../../common/notebookEditorInput.js';
 import { INotebookEditor, INotebookEditorCreationOptions } from '../notebookBrowser.js';
 import { Event } from '../../../../../base/common/event.js';
 import { Dimension } from '../../../../../base/browser/dom.js';
@@ -21,7 +20,7 @@ export interface IBorrowValue<T> {
 export interface INotebookEditorService {
 	_serviceBrand: undefined;
 
-	retrieveWidget(accessor: ServicesAccessor, groupId: number, input: NotebookEditorInput, creationOptions?: INotebookEditorCreationOptions, dimension?: Dimension, codeWindow?: CodeWindow): IBorrowValue<INotebookEditor>;
+	retrieveWidget(accessor: ServicesAccessor, groupId: number, input: { resource: URI; typeId: string }, creationOptions?: INotebookEditorCreationOptions, dimension?: Dimension, codeWindow?: CodeWindow): IBorrowValue<INotebookEditor>;
 
 	retrieveExistingWidgetFromURI(resource: URI): IBorrowValue<NotebookEditorWidget> | undefined;
 	retrieveAllExistingWidgets(): IBorrowValue<NotebookEditorWidget>[];
